[NDPE-65] Improve server start performance in production mode
start.js is intended to run in development mode as it automatically performs all
necessary build and watch steps to keep the development environment smooth.

In production, however, the files don't need to be watched and the assets don't need
to be recompiled each time the server starts. Doing so introduces a significant penalty
to startup performance on Azure especially.

Introduce separate start and postinstall scripts for development and production modes.
Development mode runs start.js as usual. In production mode, the assets are built
post-install and server.js is run directly.