NDPE-65 Support deployment to Azure (#31)
* [NDPE-65] Azure: package,json changes to support publication

Uses latest version of Node.js but a backlevel version of npm (5.0.3) to avoid errors
during the npm install portion of deployment.

Adds gulp as a direct dependency as it is used directly by the app.

* [NDPE-65] Azure: spawn locally installed gulp

A more general bugfix, probably only for Windows, but may affect systems where gulp
is not globally installed.

* [NDPE-65] Azure: use process.env.PORT directly if it is non-numeric

The port scanner function needs a numeric PORT.

* [NDPE-65] Azure: support AUTH_* environment variables to supply credentials for basic authentication

Windows pre-defines USERNAME in the environment and the dotenv npm package will not
overwrite pre-existing variable with the .env file. Therefore, to support deployment on Azure,
add support for less generic variable names that can be set via the Azure console.

* [NDPE-64] (docs) Update documentation for Azure publication and new basic auth env vars

* [NDPE-65] Improve server start performance in production mode

start.js is intended to run in development mode as it automatically performs all
necessary build and watch steps to keep the development environment smooth.

In production, however, the files don't need to be watched and the assets don't need
to be recompiled each time the server starts. Doing so introduces a significant penalty
to startup performance on Azure especially.

Introduce separate start and postinstall scripts for development and production modes.
Development mode runs start.js as usual. In production mode, the assets are built
post-install and server.js is run directly.

* [NDPE-65] (docs) Clarify Azure publication instructions with screenshot